Fixed bug in matching code; edited sample files.
Prior to this commit we were using re.finditer to find
all matches and then perform subsequent substitutions.

The problem with this is that the start/end properties
of the match objects are determined before any modifications
have been made to the underlying text.  Modifying the text
invalidates all but the first match offsets.

So now we just use re.sub and hand in a callback since
the substitution str is a function of the match.

Also, set textmode=True default var, which means that by
default input to the tagger is plain text.