Fixed bug; added basic support for lexicon regexes
The bug was just referencing option variables without
a self receiver.

The basic regex literal support is as follows:
    If a lexical entry begins with the string 'r:'
    then that entry is treated as a regex literal.

No attempt is made to allow for optional ws or tags in the
regex.  The only modification made to it is to surround it
with capture parens if they are not already present.